refactor: moved `deleteFeedData` to content-items
