+++ src/content-items/handler.ts
@@ -2,12 +2,12 @@ import { Connection } from '../connection/connection'
 import { utils } from 'ethers'
 import { Reference, RequestOptions } from '@ethersphere/bee-js'
 import { getUnixTimestamp } from '../utils/time'
-import { deleteFeedData, writeFeedData } from '../feed/api'
+import { writeFeedData } from '../feed/api'
 import { getRawDirectoryMetadataBytes } from '../directory/adapter'
 import { DIRECTORY_TOKEN, FILE_TOKEN } from '../file/handler'
 import { assertRawDirectoryMetadata, combine, splitPath } from '../directory/utils'
 import { RawDirectoryMetadata } from '../pod/types'
-import { getPathInfo, getRawMetadata } from './utils'
+import { deleteFeedData, getPathInfo, getRawMetadata } from './utils'
 import { RawMetadataWithEpoch } from './types'
 import { prepareEthAddress } from '../utils/wallet'
 import { PodPasswordBytes } from '../utils/encryption'
